LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/replaceState-inside-back-handler-expected.txt

CONSOLE MESSAGE: RangeError: Maximum call stack size exceeded.

Harness Error (FAIL), message = RangeError: Maximum call stack size exceeded.

PASS replaceState inside a navigate event for navigation.back()
